rename: postData.ts->journeyData.ts 파일명 변경
동행 관련 타입들을 선언해놓은 파일 postData.ts를 journeyData.ts로 이름 변경했습니다.
